+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/checkers/InjectConstructorChecker.kt
@@ -5,11 +5,11 @@ package dev.zacsweers.metro.compiler.fir.checkers
 import dev.zacsweers.metro.compiler.Symbols.DaggerSymbols
 import dev.zacsweers.metro.compiler.fir.MetroDiagnostics
 import dev.zacsweers.metro.compiler.fir.annotationsIn
-import dev.zacsweers.metro.compiler.fir.asFirContextualTypeKey
 import dev.zacsweers.metro.compiler.fir.classIds
 import dev.zacsweers.metro.compiler.fir.findInjectConstructor
 import dev.zacsweers.metro.compiler.fir.isAnnotatedWithAny
 import dev.zacsweers.metro.compiler.fir.validateInjectedClass
+import dev.zacsweers.metro.compiler.fir.validateInjectionSiteType
 import org.jetbrains.kotlin.diagnostics.DiagnosticReporter
 import org.jetbrains.kotlin.diagnostics.reportOn
 import org.jetbrains.kotlin.fir.analysis.checkers.MppCheckerKind
@@ -18,7 +18,6 @@ import org.jetbrains.kotlin.fir.analysis.checkers.declaration.FirClassChecker
 import org.jetbrains.kotlin.fir.declarations.FirClass
 import org.jetbrains.kotlin.fir.declarations.getAnnotationByClassId
 import org.jetbrains.kotlin.fir.declarations.primaryConstructorIfAny
-import org.jetbrains.kotlin.fir.resolve.toClassSymbol
 
 internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common) {
 
@@ -63,20 +62,14 @@ internal object InjectConstructorChecker : FirClassChecker(MppCheckerKind.Common
 
     for (parameter in constructorToValidate.valueParameterSymbols) {
       if (parameter.isAnnotatedWithAny(session, classIds.assistedAnnotations)) continue
-      val type = parameter.resolvedReturnTypeRef.coneType
-      val contextKey = type.asFirContextualTypeKey(session, null, false)
-      if (contextKey.isWrappedInLazy) {
-        val canonicalType = contextKey.typeKey.type
-        val canonicalClass = canonicalType.toClassSymbol(session) ?: continue
-        if (canonicalClass.isAnnotatedWithAny(session, classIds.assistedFactoryAnnotations)) {
-          reporter.reportOn(
-            parameter.resolvedReturnTypeRef.source ?: parameter.source ?: source,
-            MetroDiagnostics.ASSISTED_FACTORIES_CANNOT_BE_LAZY,
-            canonicalClass.name.asString(),
-            canonicalClass.classId.asFqNameString(),
-          )
-          return
-        }
+      if (
+        validateInjectionSiteType(
+          session,
+          parameter.resolvedReturnTypeRef,
+          parameter.source ?: source,
+        )
+      ) {
+        return
       }
     }
   }