+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 **Unreleased**
 --------------
 
+- **Enhancement**: Always check for available assisted factories when reporting `InvalidBinding` errors about misused assisted injects.
+- **Enhancement**: Always specifically report mismatched assisted parameter mismatches.
+- **Enhancement**: Validate `Lazy` assisted factory injections in more places.
+- **Fix**: Register `MetroDiagnostics` in FIR.
+
 0.6.6
 -----